DeckChallenge/requirements.txt
playwright>=1.45.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
opencv-python
openai>=1.3.3
